requirements.txt
bcrypt==4.0.1
fastapi==0.103.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
uvicorn==0.23.2
PyJWT==2.8.0
python-decouple==3.8